+++ CHANGELOG.md
@@ -15,9 +15,14 @@ Devemos adicionar o comentário `<!-- N/A -->` (Não adicionar), para que não p
 Caso adicionado no escopo inicial, todos os conteúdos abaixo não serão adicionados. Caso adicionado na linha, será considerado apenas ela.
 
 ## Não publicado
+### Adicioando
+-`QasSelectListDialog`:
+  - Adicionado model chamado `select-list-model` com os itens selecionados dentro do dialog.
+  - Adicionado o método `toggleDialog` no define expose.
+
 ### Corrigido
 - `QasActions`: Corrigido tamanho do botão quando passado via props e utilizado junto com `useEqualWidth`.
-- `QasSelectListDialgo`:
+- `QasSelectListDialog`:
   - Corrigido slots do dialog que não funcionavam.
   - Corrigido alinhamento do botão de adicionar item quando não se tem label no header.